Draft: move functions to draftgeoutils.offsets
